bug fix, refactor
